# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enerji Yönetim Sistemi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis/regression:
    get:
      operationId: getRegressionAnalysis
      tags: [analysis]
      summary: Regresyon analizi (EnPG, EnRÇ, EEI hesapla)
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: meterId
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegressionResult"
  /analysis/performance:
    get:
      operationId: getPerformanceIndicators
      tags: [analysis]
      summary: Enerji performans göstergeleri özeti
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceIndicators"

  # ── SWOT ──────────────────────────────────────────────────
...
  /dashboard/kpi:
    get:
      operationId: getDashboardKpi
      tags: [dashboard]
      summary: KPI özet verileri (Toplam Enerji, CO₂, TEP)
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardKpi"
  /dashboard/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [dashboard]
      summary: Aylık enerji tüketim trendi
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrendPoint"
  /dashboard/seu-breakdown:
    get:
      operationId: getSeuBreakdown
      tags: [dashboard]
      summary: ÖEK bazında enerji dağılımı
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeuBreakdownPoint"

  # ── SUMMARY ───────────────────────────────────────────────
...
components:
  schemas:
...
    RegressionResult:
      type: object
      required: [slope, intercept, r2, enpg, enrc, eei, dataPoints]
      properties:
        slope:
          type: number
        intercept:
          type: number
        r2:
          type: number
          description: "R² değeri (0-1)"
        enpg:
          type: number
          description: "Enerji Performans Göstergesi"
        enrc:
          type: number
          description: "Enerji Referans Çizgisi"
        eei:
          type: number
          description: "Enerji Etkinlik Indeksi"
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/RegressionDataPoint"
    RegressionDataPoint:
      type: object
      required: [month, actual, predicted, hdd]
      properties:
        month:
          type: integer
        actual:
          type: number
        predicted:
          type: number
        hdd:
          type: number
    PerformanceIndicators:
      type: object
      required: [totalKwh, totalTep, totalCo2, enpg, enrc, eei, savingsKwh, savingsTep, improvementPercent]
      properties:
        totalKwh:
          type: number
        totalTep:
          type: number
        totalCo2:
          type: number
        enpg:
          type: number
        enrc:
          type: number
        eei:
          type: number
        savingsKwh:
          type: number
        savingsTep:
          type: number
        improvementPercent:
          type: number

    # ── SWOT ─────────────────────────────────────────────────
...
    DashboardKpi:
      type: object
      required: [year, totalKwh, totalTep, totalCo2, kwhChange, tepChange, co2Change, meterCount, activeSeuCount]
      properties:
        year:
          type: integer
        totalKwh:
          type: number
        totalTep:
          type: number
        totalCo2:
          type: number
        kwhChange:
          type: number
        tepChange:
          type: number
        co2Change:
          type: number
        meterCount:
          type: integer
        activeSeuCount:
          type: integer
    MonthlyTrendPoint:
      type: object
      required: [month, monthName, kwh, tep, co2]
      properties:
        month:
          type: integer
        monthName:
          type: string
        kwh:
          type: number
        tep:
          type: number
        co2:
          type: number
        hdd:
          type: ["number", "null"]
        cdd:
          type: ["number", "null"]
    SeuBreakdownPoint:
      type: object
      required: [name, kwh, percentage, category]
      properties:
        name:
          type: string
        kwh:
          type: number
        percentage:
          type: number
        category:
          type: string

    # ── Summary ───────────────────────────────────────────────
...